requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langgraph>=0.0.10
qdrant-client>=1.7.0
sentence-transformers>=2.2.2
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
pytest>=7.4.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
python-magic>=0.4.27
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
openai>=1.0.0
aiohttp>=3.8.0
requests>=2.31.0
beautifulsoup4>=4.12.0 